Initial refactor of snackbar into GameInfo component.
