+++ pilgrims-client/src/components/Game.vue
@@ -3,14 +3,7 @@
     <v-layout column fill-height>
       <v-layout row class="main-area">
         <v-flex xs9 @mousemove="mouseOver">
-          <v-snackbar style="text-align: left"
-            v-model="showGameInfo"
-            :timeout="5000"
-            :top="'top'">
-          <small v-if="gameID">Tell your friends to join this game at: <b>{{gameID}}</b></small>
-          <small v-if="currentPlayer" class="pad">Current player:<b>{{currentPlayer.name}}</b></small>
-          <small v-if="currentDie" class="pad">Current die: <b>{{currentDie.toString()}}</b></small>
-          </v-snackbar>
+          <GameInfo v-bind="showGameInfo"></GameInfo>
           <Map/>
         </v-flex>
         <v-flex xs3>
@@ -31,6 +24,7 @@ import Map from './Map.vue';
 import Player from './Player.vue';
 import CustomizeGame from './CustomizeGame.vue';
 import Error from './Error.vue';
+import GameInfo from './GameInfo.vue';
 
 @Component({
   components: {
@@ -39,6 +33,7 @@ import Error from './Error.vue';
     Player,
     CustomizeGame,
     Error,
+    GameInfo,
   },
 })
 export default class Game extends Vue {
@@ -47,19 +42,11 @@ export default class Game extends Vue {
   get started() {
     return this.$store.getters['game/getIsGameStarted'];
   }
-  get gameID() {
-    return this.$store.state.game.gameId;
-  }
-  get currentPlayer() {
-    return this.$store.getters['game/getCurrentPlayer'];
-  }
-  get currentDie() {
-    return this.$store.getters['game/getCurrentDie'];
-  }
 
   public mouseOver(e: MouseEvent) {
     if (e.offsetY <= 70) {
-      if (e) this.showGameInfo = true;
+      debugger;
+      this.showGameInfo = true;
     } else {
       this.showGameInfo = false;
     }
